+++ app/channels/game_channel.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class GameChannel < ApplicationCable::Channel
+  def subscribed
+    stream_from "game-#{params[:game_id]}"
+  end
+end